fix: extract message_query from storage.rs (2500-LOC ceiling)
Move get_thread, collect_thread_messages, and find_message to
storage/message_query.rs. Brings storage.rs from 2522 to 2459 LOC.


Agend-Agent: claude-aef7c0
Agend-Task: t-20260715060519762645-64650-116
Agend-Branch: fix/ci-handoff-episode-settlement
Agend-Issued-At: 2026-07-15T06:05:29.565132+00:00
